cleaned up code slightly
