+++ public/login.html
@@ -9,7 +9,7 @@
 </head>
 <body>
     <div class = "container">
-        <form method="POST" action ="/log">
+        <form method="POST" action ="/login">
             <label for="username"><b>Username</b></label>
             <input type="text" placeholder="Enter Username" name="username" required>
             <br>
