--- public/registration.html
@@ -1,73 +1,0 @@
-<!DOCTYPE html>
-<html>
-<head>
-    <title>Registration Form</title>
-    
-</head>
-<body>
-    <div class = "container">
-        <form>
-            <label for="username"><b>Username</b></label>
-            <input type="text" placeholder="Enter Username" name="username" required>
-            <br>
-            <label for="first_name"><b>First Name</b></label>
-            <input type="text" placeholder="Enter your first name" name="first_name" required>
-            <br>
-            <label for="last_name"><b>Last Name</b></label>
-            <input type="text" placeholder="Enter your last name" name="last_name" required>
-            <br>
-            <label for="email"><b>Email</b></label>
-            <input type="email" placeholder="Enter Email" name="email" required>
-            <br>
-            <label for="password"><b>Password</b></label>
-            <input class="pw" type="password" placeholder="Enter Password" name="password" required>
-            <br>
-            <label for="password_reenter"><b>Re-Enter Password</b></label>
-            <input class="repw" type="password" placeholder="Enter Password" name="password_reenter" required>
-            <br>
-            <button type="submit">Create Account!</button>
-
-
-        </form>
-    </div>
-    <script src="https://ajax.googleapis.com/ajax/libs/jquery/3.3.1/jquery.min.js"></script>
-    <script type="text/javascript">
-      //on click of submit button, javascript checks whether password is between 8 - 16 characters long
-        $("button").on("click",function(event){
-            event.preventDefault();
-            var PW = $(".pw").val();
-            var RePW = $(".repw").val();
-            if(PW !== RePW)
-            {
-                alert("Passwords do not match");
-                $(".pw").val("");
-                $(".repw").val("");
-            }
-            else
-            {
-                var PWL = $(".pw").val().length;
-                // alert(PWL);
-                if(PWL < 8 || PWL > 16)
-                {
-                    alert("Password must be between 8 to 16 characters long");
-                    $(".pw").val("");
-                    $(".repw").val("");
-                }
-                else
-                {
-                    $("form").attr("method","POST");
-                    $("form").attr("action","/signup");
-                    $("form").submit();
-                }
-                console.log(PWL);
-            }
-         
-        })
-
- 
-    </script>>
-
-   
-
-</body>
-</html>
