+++ views/pages/registration.ejs
@@ -60,7 +60,7 @@
                 else
                 {
                     $("form").attr("method","POST");
-                    $("form").attr("action","/register");
+                    $("form").attr("action","/signup");
                     $("form").submit();
                 }
                 console.log(PWL);
